+++ vsd-mp4/src/decrypt/mod.rs
@@ -2,7 +2,8 @@ mod decrypter;
 mod error;
 mod hls;
 mod processor;
+pub mod stream;
 
 pub use error::{DecryptError, Result};
 pub use hls::{HlsAes128Decrypter, HlsSampleAesDecrypter};
-pub use processor::{CencDecryptingProcessor, CencDecryptingProcessorBuilder, DecryptionSession};
+pub use processor::{CencDecryptingProcessor, CencDecryptingProcessorBuilder};